Refactored files, reduced package size
