+++ config.js
@@ -2,7 +2,6 @@
 
 var path = require( 'path' );
 
-
 module.exports = {
   filePaths : {
     analyzers : path.join( __dirname, 'lib/analyzers' )
