+++ lib/analyzers.js
