+++ lib/analyzers/jspm/index.js
@@ -2,8 +2,9 @@
 
 var fs          = require( 'fs' );
 var path        = require( 'path' );
-var creditUtil  = require( '../../../util/credit' );
-var packageUtil = require( '../../../util/package' );
+
+var creditUtil  = require( '../../credit' );
+var packageUtil = require( '../../package' );
 
 
 /**
