+++ lib/credit.js
