+++ lib/package.js
