+++ package.json
@@ -3,9 +3,14 @@
   "version": "1.1.0",
   "description": "Find out on whose work your project is based on",
   "main": "index.js",
+  "files": [
+    "config.js",
+    "index.js",
+    "lib"
+  ],
   "scripts": {
-    "lint": "eslint *.js util/* lib/*",
-    "test": "npm run lint && nyc ava *.spec.js ./util/*.spec.js ./lib/**/*.spec.js",
+    "lint": "eslint lib/*.js lib/**/*.js",
+    "test": "npm run lint && nyc ava",
     "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
@@ -22,6 +27,7 @@
     "ava": "^0.16.0",
     "coveralls": "^2.11.4",
     "eslint": "^3.7.1",
+    "fs-extra": "^0.30.0",
     "nyc": "^8.3.1",
     "tmp": "0.0.29"
   },
