+++ test/fixtures/jspm_packages/github/components/foo/package.json
@@ -1,0 +1,1 @@
+{"author":"Don Calsow"}
