+++ test/fixtures/jspm_packages/npm/baz/package.json
@@ -1,0 +1,1 @@
+{"author":"Alice Bobson","maintainers":["Randy Ran",{"name":"Bobby Bob","email":"bobby@bob.io"}]}
