+++ test/fixtures/jspm_packages/npm/boing/package.json
@@ -1,0 +1,1 @@
+{"author":"Bob Calsow <bob@calsow.io>"}
