+++ test/fixtures/jspm_packages/npm/boo/package.json
@@ -1,0 +1,1 @@
+{"author":"Janice Robson"}
