+++ test/fixtures/linked/package.json
@@ -1,0 +1,1 @@
+{"author":"Bob Loblaw"}
